* Rename COMPRESSIONMETRIC -> CompressionMetric * Refactor compression_analysis into separate functions for metric calculation and plotting + analysis * Add docstrings for modules * Fix linting errors
