+++ subcell_pipeline/analysis/compression_metrics/constants.py
@@ -1,3 +1,5 @@
+# Constants used in compression metric analysis.
+
 ABSOLUTE_TOLERANCE: float = 1e-6
 """
 The absolute tolerance for vector length.
@@ -8,5 +10,5 @@ SIMULATOR_COLOR_MAP: dict[str, str] = {
     "cytosim": "#008080",
 }
 """Map of simulator name to color."""
-DEFAULT_BENDING_CONSTANT: float = 1
+DEFAULT_BENDING_CONSTANT: float = 1.0
 """The default bending constant in pN nm."""
